resources/python/sources/snowplow_source/requirements.txt
quixstreams==2.9.0
boto3
snowplow_analytics_sdk
python-dotenv